AST-5022 match GitHub validator structure - rules and results are now parsed in a single loop - all the result's nodes are added as a location to the result
Notes:
- a workaround for query id due to AST returning it in different types depending on engine is present
- when the rule id is a CVE, we don't have a friendly name available
- helpUri for each rule is AST docs
- friendly name, when existing, probably won't match PascalCase
